+++ R/discomb.R
@@ -507,8 +507,11 @@ discomb <- function(TE, seTE,
     else if (!compmatch(labels, "*"))
       sep.trts <- "*"
     else
-      stop("All predefined separators (':', '-', '_', '/', '+', '.', '|', '*') are used in at least one treatment label.",
-           "\n   Please specify a different character that should be used as separator (argument 'sep.trts').",
+      stop("All predefined separators ",
+           "(':', '-', '_', '/', '+', '.', '|', '*') ",
+           "are used in at least one treatment label.",
+           "\n   Please specify a different character that ",
+           "should be used as separator (argument 'sep.trts').",
            call. = FALSE)
   }
   ##
@@ -701,19 +704,39 @@ discomb <- function(TE, seTE,
   colnames(X.matrix) <- colnames(C.matrix)
   rownames(X.matrix) <- studlab
   ##
-  sel.trts <- character(0)
+  sel.comps <- character(0)
   ##
   if (qr(X.matrix)$rank < c) {
     sum.trts <- apply(abs(X.matrix), 2, sum)
-    sel.trts <- names(sum.trts)[sum.trts == 0]
+    sel.comps <- gsub("^\\s+|\\s+$", "",
+                      names(sum.trts)[sum.trts == 0])
     ##
-    if (length(sel.trts > 0))
+    if (length(sel.comps > 0)) {
+      list.trts <- lapply(compsplit(labels, sep.comps),
+                          gsub, pattern = "^\\s+|\\s+$", replacement = "")
+      sel.combs <- rep(NA, length(list.trts))
+      ##
+      for (i in seq_along(list.trts))
+        sel.combs[i] <- any(list.trts[[i]] %in% sel.comps)
+      ##
+      sel.combs <- labels[sel.combs]
+      ##
       warning("The following treatment component",
-              if (length(sel.trts) > 1) "s",
+              if (length(sel.comps) > 1) "s",
               " cannot be estimated: ",
-               paste(paste("'", sel.trts, "'", sep = ""),
+               paste(paste("'", sel.comps, "'", sep = ""),
                      collapse = ", "),
-              call. = FALSE)
+              "\nAccordingly, the following treatment combination",
+              if (length(sel.combs) > 1) "s",
+             " cannot be estimated:\n",
+               paste(paste("'", sel.combs, "'", sep = ""),
+                     collapse = ", "),
+             "\nPlease consider whether a network meta-analysis without",
+             " inestimable component",
+             if (length(sel.comps) > 1) "s",
+             " is sensible.",
+             call. = FALSE)
+    }
     else
       warning("A total of ", c - qr(X.matrix)$rank, " of ", c,
               " treatment components cannot be estimated.",
@@ -990,87 +1013,83 @@ discomb <- function(TE, seTE,
   ##
   ## Remove estimates for inestimable combinations and components
   ##
-  if (length(sel.trts) > 0) {
+  if (length(sel.comps) > 0) {
     ##
-    res$c <- res$c - length(sel.trts)
+    res$c <- res$c - length(sel.comps)
     ##
     ## Identify combinations
     ##
-    sellist <- compsplit(res$trts, sep.comps)
-    sel.i <- rep(NA, length(sellist))
+    list.trts <- lapply(compsplit(res$trts, sep.comps),
+                        gsub, pattern = "^\\s+|\\s+$", replacement = "")
+    sel1 <- rep(NA, length(list.trts))
     ##
-    sel.trts <- meta:::rmSpace(meta:::rmSpace(sel.trts), end = TRUE)
+    for (i in seq_along(list.trts))
+      sel1[i] <- any(list.trts[[i]] %in% sel.comps)
     ##
-    for (i in seq_along(sellist))
-      sel.i[i] <- any(meta:::rmSpace(meta:::rmSpace(sellist[[i]]),
-                                     end = TRUE) %in% sel.trts)
+    res$Comb.fixed[sel1] <- NA
+    res$seComb.fixed[sel1] <- NA
+    res$lower.Comb.fixed[sel1] <- NA
+    res$upper.Comb.fixed[sel1] <- NA
+    res$statistic.Comb.fixed[sel1] <- NA
+    res$pval.Comb.fixed[sel1] <- NA
     ##
-    res$Comb.fixed[sel.i] <- NA
-    res$seComb.fixed[sel.i] <- NA
-    res$lower.Comb.fixed[sel.i] <- NA
-    res$upper.Comb.fixed[sel.i] <- NA
-    res$statistic.Comb.fixed[sel.i] <- NA
-    res$pval.Comb.fixed[sel.i] <- NA
+    res$Comb.random[sel1] <- NA
+    res$seComb.random[sel1] <- NA
+    res$lower.Comb.random[sel1] <- NA
+    res$upper.Comb.random[sel1] <- NA
+    res$statistic.Comb.random[sel1] <- NA
+    res$pval.Comb.random[sel1] <- NA
     ##
-    res$Comb.random[sel.i] <- NA
-    res$seComb.random[sel.i] <- NA
-    res$lower.Comb.random[sel.i] <- NA
-    res$upper.Comb.random[sel.i] <- NA
-    res$statistic.Comb.random[sel.i] <- NA
-    res$pval.Comb.random[sel.i] <- NA
+    res$TE.fixed[sel1, ] <- NA
+    res$seTE.fixed[sel1, ] <- NA
+    res$lower.fixed[sel1, ] <- NA
+    res$upper.fixed[sel1, ] <- NA
+    res$statistic.fixed[sel1, ] <- NA
+    res$pval.fixed[sel1, ] <- NA
     ##
-    res$TE.fixed[sel.i, ] <- NA
-    res$seTE.fixed[sel.i, ] <- NA
-    res$lower.fixed[sel.i, ] <- NA
-    res$upper.fixed[sel.i, ] <- NA
-    res$statistic.fixed[sel.i, ] <- NA
-    res$pval.fixed[sel.i, ] <- NA
+    res$TE.fixed[, sel1] <- NA
+    res$seTE.fixed[, sel1] <- NA
+    res$lower.fixed[, sel1] <- NA
+    res$upper.fixed[, sel1] <- NA
+    res$statistic.fixed[, sel1] <- NA
+    res$pval.fixed[, sel1] <- NA
     ##
-    res$TE.fixed[, sel.i] <- NA
-    res$seTE.fixed[, sel.i] <- NA
-    res$lower.fixed[, sel.i] <- NA
-    res$upper.fixed[, sel.i] <- NA
-    res$statistic.fixed[, sel.i] <- NA
-    res$pval.fixed[, sel.i] <- NA
+    res$TE.random[sel1, ] <- NA
+    res$seTE.random[sel1, ] <- NA
+    res$lower.random[sel1, ] <- NA
+    res$upper.random[sel1, ] <- NA
+    res$statistic.random[sel1, ] <- NA
+    res$pval.random[sel1, ] <- NA
     ##
-    res$TE.random[sel.i, ] <- NA
-    res$seTE.random[sel.i, ] <- NA
-    res$lower.random[sel.i, ] <- NA
-    res$upper.random[sel.i, ] <- NA
-    res$statistic.random[sel.i, ] <- NA
-    res$pval.random[sel.i, ] <- NA
-    ##
-    res$TE.random[, sel.i] <- NA
-    res$seTE.random[, sel.i] <- NA
-    res$lower.random[, sel.i] <- NA
-    res$upper.random[, sel.i] <- NA
-    res$statistic.random[, sel.i] <- NA
-    res$pval.random[, sel.i] <- NA
+    res$TE.random[, sel1] <- NA
+    res$seTE.random[, sel1] <- NA
+    res$lower.random[, sel1] <- NA
+    res$upper.random[, sel1] <- NA
+    res$statistic.random[, sel1] <- NA
+    res$pval.random[, sel1] <- NA
     ##
     ## Identify components
     ##
-    sellist <- compsplit(res$comps, sep.comps)
-    sel.i <- rep(NA, length(sellist))
+    list.comps <- lapply(compsplit(res$comps, sep.comps),
+                         gsub, pattern = "^\\s+|\\s+$", replacement = "")
+    sel2 <- rep(NA, length(list.comps))
     ##
-    sel.trts <- meta:::rmSpace(meta:::rmSpace(sel.trts), end = TRUE)
+    for (i in seq_along(list.comps))
+      sel2[i] <- any(list.comps[[i]] %in% sel.comps)
     ##
-    for (i in seq_along(sellist))
-      sel.i[i] <- any(meta:::rmSpace(meta:::rmSpace(sellist[[i]]),
-                                     end = TRUE) %in% sel.trts)
+    res$Comp.fixed[sel2] <- NA
+    res$seComp.fixed[sel2] <- NA
+    res$lower.Comp.fixed[sel2] <- NA
+    res$upper.Comp.fixed[sel2] <- NA
+    res$statistic.Comp.fixed[sel2] <- NA
+    res$pval.Comp.fixed[sel2] <- NA
     ##
-    res$Comp.fixed[sel.i] <- NA
-    res$seComp.fixed[sel.i] <- NA
-    res$lower.Comp.fixed[sel.i] <- NA
-    res$upper.Comp.fixed[sel.i] <- NA
-    res$statistic.Comp.fixed[sel.i] <- NA
-    res$pval.Comp.fixed[sel.i] <- NA
-    ##
-    res$Comp.random[sel.i] <- NA
-    res$seComp.random[sel.i] <- NA
-    res$lower.Comp.random[sel.i] <- NA
-    res$upper.Comp.random[sel.i] <- NA
-    res$statistic.Comp.random[sel.i] <- NA
-    res$pval.Comp.random[sel.i] <- NA
+    res$Comp.random[sel2] <- NA
+    res$seComp.random[sel2] <- NA
+    res$lower.Comp.random[sel2] <- NA
+    res$upper.Comp.random[sel2] <- NA
+    res$statistic.Comp.random[sel2] <- NA
+    res$pval.Comp.random[sel2] <- NA
   }
   
   res